2022.11.18

+++ src/algorithm_study/부족한금액계산하기/Solution.java
@@ -5,7 +5,7 @@ public class Solution {
 	public long solution(int price, int money, int count) {
 		long answer = 0;
 		for (int i = 1; i <= count; i++) {
-			answer += price * i;
+			answer += (long) price * i;
 		}
 		answer = money - answer;
